@@ -1,18 +1,16 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import 'normalize.css'
-import './index.css'
-import App from './App'
-import { AppProvider } from './context/appContext'
-ReactDOM.render(
-  <React.StrictMode>
-    <AppProvider>
-      <App />
-    </AppProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import 'normalize.css';
+import './index.css';
+import App from './App';
+import { store } from './store';
+import { Provider } from 'react-redux';
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
+  <Provider store={store}>
+    <App tab='home' />
+  </Provider>
+);
