@@ -1,30 +1,18 @@
-import React from 'react';
-
-const FormRow = ({
-  type,
-  name,
-  value,
-  handleChange,
-  horizontal,
-  placeholder,
-}) => {
+const FormRow = ({ type, name, value, handleChange, labelText }) => {
   return (
     <div className='form-row'>
-      {!horizontal && (
-        <label htmlFor={name} className='form-label'>
-          {name}
-        </label>
-      )}
+      <label htmlFor={name} className='form-label'>
+        {labelText || name}
+      </label>
       <input
+        id={name}
         type={type}
-        value={value}
         name={name}
+        value={value}
         onChange={handleChange}
         className='form-input'
-        placeholder={placeholder}
       />
     </div>
   );
 };
-
 export default FormRow;
