@@ -1,7 +1,6 @@
-import Register from './Register'
-import Home from './Home'
-import Dashboard from './Dashboard'
-import Edit from './Edit'
-import Error from './Error'
-import PrivateRoute from './PrivateRoute'
-export { Home, Register, Dashboard, Edit, Error, PrivateRoute }
+import Error from './Error';
+import Landing from './Landing';
+import ProtectedRoute from './ProtectedRoute';
+import Register from './Register';
+
+export { Landing, Error, Register, ProtectedRoute };
